requirements.txt
# Kafka and Spark dependencies
kafka-python>=2.0.2
pyspark>=3.2.0

# Database dependencies
mysql-connector-python>=8.0.26
sqlalchemy>=1.4.23
pandas>=1.3.3

# Visualization dependencies
matplotlib>=3.4.3
seaborn>=0.11.2

# Utility dependencies
python-dotenv>=0.19.0
loguru>=0.5.3